icu_provider::baked

Module zerotrie

Data stored as as ZeroTrieSimpleAscii

Re-exports

pub use crate::DynamicDataMarker;

Structs

Data

Regular baked data: a trie for lookups and a slice of values

DataForVarULEs

Optimized data stored as a single VarZeroSlice to reduce token count

DataRef

Regular baked data: a trie for lookups and a slice of values

ZeroTrieSimpleAscii

A data structure that compactly maps from ASCII strings to integers.

Constants

ID_SEPARATOR

This is a valid separator as DataLocale will never produce it.