icu_provider

Module data_provider

Structs

DataProviderWithMarker

A DataProvider associated with a specific marker.

Traits

BoundDataProvider

A data provider that loads data for a specific data type.

DataProvider

A data provider that loads data for a specific DataMarkerInfo.

DryDataProvider

A data provider that can determine whether it can load a particular data identifier, potentially cheaper than actually performing the load.

DynamicDataProvider

A data provider that loads data for a specific data type.

DynamicDryDataProvider

A dynanmic data provider that can determine whether it can load a particular data identifier, potentially cheaper than actually performing the load.

IterableDataProvider

A DataProvider that can iterate over all supported DataIdentifierCows.

IterableDynamicDataProvider

A DynamicDataProvider that can iterate over all supported DataIdentifierCows for a certain marker.