Trait icu_provider::data_provider::DataProvider

pub trait DataProvider<'d, T: ?Sized> where
    T: ToOwned,
    <T as ToOwned>::Owned: Debug, {
    fn load_payload(
        &self, 
        req: &DataRequest
    ) -> Result<DataResponse<'d, T>, Error>;
}

A generic data provider that loads a payload of a specific type.

Required methods

fn load_payload(&self, req: &DataRequest) -> Result<DataResponse<'d, T>, Error>

Query the provider for data, returning the result.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.

Loading content...