Module icu_provider::datagen

This module contains various utilities required to generate ICU4X data files, typically via the icu4x-datagen tool. End users should not need to consume anything in this module as a library unless defining new types that integrate with icu4x-datagen.

This module can be enabled with the datagen feature on icu4x-provider-core.

Structs

HeapStats

Stats on the heap size needed when attempting to zero-copy-deserialize a postcard-formatted data struct.

HeapStatsMarker

The DataMarker marker type for HeapStats.

ReturnedPayloadError

Traits

DataConverter

A trait that allows for converting between data payloads of different types.

IterableDynProvider

A DynProvider that can iterate over all supported ResourceOptions for a certain key.

IterableResourceProvider

A ResourceProvider that can iterate over all supported ResourceOptions for a certain key.

OmnibusDatagenProvider

A blanket-implemented trait that allows for all the datagen-relevant traits to be used in trait objects. The type parameter is the marker type needed for IterableDynProvider.