Trait icu_provider::datagen::IterableDynProvider

pub trait IterableDynProvider<M: DataMarker>: DynProvider<M> {
    fn supported_options_for_key(
        &self, 
        key: ResourceKey
    ) -> Result<Box<dyn Iterator<Item = ResourceOptions>>, DataError>;
}

A DynProvider that can iterate over all supported ResourceOptions for a certain key.

Implementing this trait means that a data provider knows all of the data it can successfully return from a load request.

Required methods

fn supported_options_for_key(
&self,
key: ResourceKey
) -> Result<Box<dyn Iterator<Item = ResourceOptions>>, DataError>

Given a ResourceKey, returns a boxed iterator over ResourceOptions.