icu_provider

Module dynutil

Utilities for using trait objects with DataPayload.

Macros

impl_casting_upcast

Implements UpcastDataPayload from several data markers to a single data marker that all share the same DynamicDataMarker::DataStruct.

impl_dynamic_data_provider

Implements DynamicDataProvider for a marker type S on a type that already implements DynamicDataProvider or DataProvider for one or more M, where M is a concrete type that is convertible to S via UpcastDataPayload.

impl_iterable_dynamic_data_provider

Traits

UpcastDataPayload

Trait to allow conversion from DataPayload<T> to DataPayload<S>.