Trait icu_provider::dynutil::UpcastDataPayload

pub trait UpcastDataPayload<'d, 's, M> where
    M: DataMarker<'s>,
    Self: Sized + DataMarker<'s>, {
    fn upcast(other: DataPayload<'d, 's, M>) -> DataPayload<'d, 's, Self>;
}

Trait to allow conversion from DataPayload<T> to DataPayload<S> where T implements S.

This is used internally by impl_dyn_provider! and is not intended to be called from userland code. You may be looking for DataPayload::downcast, which converts in the other direction.

Required methods

fn upcast(other: DataPayload<'d, 's, M>) -> DataPayload<'d, 's, Self>

Upcast a DataPayload<T> to a DataPayload<S> where T implements trait S.

Examples

Upcast and then downcast a data struct of type Cow<str> (cart type String) via ErasedDataStruct:

use icu_provider::prelude::*;
use icu_provider::erased::*;
use icu_provider::dynutil::UpcastDataPayload;
use icu_provider::marker::CowStringMarker;
use std::borrow::Cow;

let data = "foo".to_string();
let original = DataPayload::<CowStringMarker>::from_owned(Cow::Owned(data));
let upcasted = ErasedDataStructMarker::upcast(original);
let downcasted = upcasted
    .downcast::<CowStringMarker>()
    .expect("Type conversion");
assert_eq!(downcasted.get(), "foo");

Loading content...