icu_provider

Enum DataErrorKind

#[non_exhaustive]
pub enum DataErrorKind {
    MarkerNotFound,
    IdentifierNotFound,
    InvalidRequest,
    InconsistentData(DataMarkerInfo),
    Downcast(&'static str),
    Deserialize,
    Custom,
}

A list specifying general categories of data provider error.

Errors may be caused either by a malformed request or by the data provider not being able to fulfill a well-formed request.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

MarkerNotFound

No data for the requested data marker. This is only returned by DynamicDataProvider.

IdentifierNotFound

There is data for the data marker, but not for this particular data identifier.

InvalidRequest

The request is invalid, such as a request for a singleton marker containing a data identifier.

InconsistentData(DataMarkerInfo)

The data for two DataMarkers is not consistent.

Downcast(&'static str)

An error occured during Any downcasting.

Deserialize

An error occured during [serde] deserialization.

Check debug logs for potentially more information.

Custom

An unspecified error occurred.

Check debug logs for potentially more information.

Implementations

impl DataErrorKind

pub const fn into_error(self) -> DataError

Converts this DataErrorKind into a DataError.

If possible, you should attach context using a with_ function.

pub const fn with_marker(self, marker: DataMarkerInfo) -> DataError

Creates a DataError with a data marker context.

pub const fn with_str_context(self, context: &'static str) -> DataError

Creates a DataError with a string context.

pub fn with_type_context<T>(self) -> DataError

Creates a DataError with a type name context.

pub fn with_req(self, marker: DataMarkerInfo, req: DataRequest<'_>) -> DataError

Creates a DataError with a request context.