Module icu_provider::erased

Re-exports

pub use erased_serde;

Structs

DataReceiver
Concrete struct backing ErasedDataReceiver. Implemented for:

Traits

ErasedDataProvider
A type-erased data provider that loads a payload of any type.

ErasedDataReceiver
An receiver capable of accepting type-erased data.

ErasedDataStruct
Auto-implemented trait allowing for type erasure of data provider structs. Requires the static lifetime in order to be convertible to Any.