Trait icu_provider::erased::ErasedDataProvider

pub trait ErasedDataProvider<'d> {
    fn load_to_receiver(
        &self, 
        req: &DataRequest, 
        receiver: &mut dyn ErasedDataReceiver<'d, '_>
    ) -> Result<DataResponseMetadata, Error>;
}

A type-erased data provider that loads a payload of any type.

Required methods

fn load_to_receiver(
&self,
req: &DataRequest,
receiver: &mut dyn ErasedDataReceiver<'d, '_>
) -> Result<DataResponseMetadata, Error>

Query the provider for data, loading it into a ErasedDataReceiver.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.

Loading content...