Trait icu_provider::erased::ErasedDataReceiver

pub trait ErasedDataReceiver<'d, 'de> {
    fn receive_deserializer(
        &mut self, 
        deserializer: &mut dyn Deserializer<'de>
    ) -> Result<(), Error>;

    fn receive_default(&mut self) -> Result<(), Error>;

    fn receive_erased(
        &mut self, 
        erased: Cow<'d, dyn ErasedDataStruct>
    ) -> Result<(), Error>;

    fn take_erased(&mut self) -> Result<Cow<'d, dyn ErasedDataStruct>, Error>;

    fn has_payload(&self) -> bool;

    fn reset(&mut self);
}

An receiver capable of accepting type-erased data.

Lifetimes:

'd = lifetime of borrowed data (Cow::Borrowed)

'de = deserializer lifetime; can usually be '_

Required methods

fn receive_deserializer(
&mut self,
deserializer: &mut dyn Deserializer<'de>
) -> Result<(), Error>

Consumes a Serde Deserializer into this ErasedDataReceiver as owned data.

This method results in an owned payload, but the payload could have non-static references according to the deserializer lifetime.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

const JSON: &'static str = "\"hello world\"";

let mut receiver = DataReceiver::<String>::new();
let mut d = serde_json::Deserializer::from_str(JSON);
receiver.receive_deserializer(&mut erased_serde::Deserializer::erase(&mut d))
    .expect("Deserialization should be successful");

assert!(matches!(receiver.payload, Some(Cow::Owned(_))));
assert_eq!("hello world", *receiver.borrow_payload().unwrap());

fn receive_default(&mut self) -> Result<(), Error>

Sets the payload to the default value. Assumes Default is implemented for the type.

This method results in an owned payload.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let mut receiver = DataReceiver::<String>::new();
receiver.receive_default().expect("Default is implemented for String");

assert!(matches!(receiver.payload, Some(Cow::Owned(_))));
assert_eq!("", *receiver.borrow_payload().unwrap());

fn receive_erased(
&mut self,
erased: Cow<'d, dyn ErasedDataStruct>
) -> Result<(), Error>

Sets the payload to the value contained in the given Cow of ErasedDataStruct. May be owned or borrowed.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let local_data = "hello world".to_string();

let mut receiver = DataReceiver::<String>::new();
receiver.receive_erased(Cow::Borrowed(&local_data)).expect("Types should match");

assert!(matches!(receiver.payload, Some(Cow::Borrowed(_))));
assert_eq!("hello world", *receiver.borrow_payload().unwrap());

fn take_erased(&mut self) -> Result<Cow<'d, dyn ErasedDataStruct>, Error>

Takes the payload as a Cow of ErasedDataStruct. Error if not present.

This method is useful if you have an ErasedDataReceiver trait object. If you have a fully typed DataReceiver, you should use take_payload() instead.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let mut receiver = DataReceiver::<String>::new();
receiver.receive_default().expect("Default is implemented for String");

assert!(receiver.has_payload());
let erased_payload = receiver.take_erased().unwrap();
assert!(!receiver.has_payload());

assert_eq!("", erased_payload.as_any().downcast_ref::<String>().unwrap());

fn has_payload(&self) -> bool

Returns whether the receiver has a payload.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let mut receiver = DataReceiver::<String>::new();
assert!(!receiver.has_payload());

receiver.receive_default().expect("Default is implemented for String");
assert!(receiver.has_payload());

fn reset(&mut self)

Discards the payload if present.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let mut receiver = DataReceiver::<String>::new();
receiver.receive_default().expect("Default is implemented for String");

assert!(receiver.has_payload());
receiver.reset();
assert!(!receiver.has_payload());

Loading content...
Implementations

impl<'a, 'd, '_> dyn ErasedDataReceiver<'d, '_> + 'a

pub fn receive_payload<T>(&mut self, payload: Cow<'d, T>) -> Result<(), Error> where
T: Serialize + Clone + Debug + Any,

Convenience method: sets the payload to the value contained in the given Cow of a concrete type. May be owned or borrowed.

This method is useful for transferring payloads between receivers.

Example

use icu_provider::prelude::*;
use icu_provider::erased::DataReceiver;
use icu_provider::erased::ErasedDataReceiver;
use std::borrow::Cow;

let mut receiver1 = DataReceiver::<f64>::new();
let mut receiver2 = DataReceiver::<f64>::new();

receiver1.receive_default();
assert!(receiver1.payload.is_some());
assert_eq!(0.0, *receiver1.borrow_payload().unwrap());

(&mut receiver2 as &mut dyn ErasedDataReceiver)
    .receive_payload(receiver1.take_payload().unwrap());

assert!(receiver1.payload.is_none());
assert_eq!(0.0, *receiver2.borrow_payload().unwrap());