Trait icu_provider::export::DataExporter

pub trait DataExporter<'data, M> where
    M: DataMarker<'data>, {
    fn put_payload(
        &mut self, 
        req: DataRequest, 
        payload: DataPayload<'data, M>
    ) -> Result<(), Error>;

    fn flush(&mut self) -> Result<(), Error> { ... }

    fn close(&mut self) -> Result<(), Error> { ... }
}

An object capable of serializing data payloads to be read by a DataProvider.

A DataProvider by itself is “read-only”; this trait enables it to be “read-write”.

Required methods

fn put_payload(
&mut self,
req: DataRequest,
payload: DataPayload<'data, M>
) -> Result<(), Error>

Save a payload corresponding to the given data request (resource path).

Provided methods

fn flush(&mut self) -> Result<(), Error>

Function called after a key has been fully dumped into the exporter.

fn close(&mut self) -> Result<(), Error>

Function called after all keys have been fully dumped.