Module icu_provider::hello_world

Data provider returning multilingual “Hello World” strings for testing.

Structs

FormattedHelloWorld

A formatted hello world message. Implements Writeable.

HelloWorldFormatter

A type that formats localized “hello world” strings.

HelloWorldJsonProvider

A data provider returning Hello World strings in different languages as JSON blobs.

HelloWorldProvider

A data provider returning Hello World strings in different languages.

HelloWorldV1

A struct containing “Hello World” in the requested language.

HelloWorldV1Marker

Marker type for HelloWorldV1.