icu_provider::hello_world

Struct HelloWorldFormatter

pub struct HelloWorldFormatter { /* private fields */ }

A type that formats localized โhello worldโ strings.

This type is intended to take the shape of a typical ICU4X formatter API.

Examples

use icu_locid::locale;
use icu_provider::hello_world::{HelloWorldFormatter, HelloWorldProvider};
use writeable::assert_writeable_eq;

let fmt = HelloWorldFormatter::try_new_unstable(
    &HelloWorldProvider,
    &locale!("eo").into(),
)
.expect("locale exists");

assert_writeable_eq!(fmt.format(), "Saluton, Mondo");

Implementations

impl HelloWorldFormatter

pub fn try_new(locale: &DataLocale) -> Result<Self, DataError>

Creates a new HelloWorldFormatter for the specified locale.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), locale: &DataLocale, ) -> Result<Self, DataError>

A version of [Self :: try_new] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_unstable<P>( provider: &P, locale: &DataLocale, ) -> Result<Self, DataError>
where P: DataProvider<HelloWorldV1Marker>,

A version of Self::try_new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn format<'l>(&'l self) -> FormattedHelloWorld<'l>

Formats a hello world message, returning a FormattedHelloWorld.

pub fn format_to_string(&self) -> String

Formats a hello world message, returning a String.