Struct icu_provider::inv::InvariantDataProvider

pub struct InvariantDataProvider;

A locale-invariant data provider. Sometimes useful for testing. Not intended to be used in production environments.

The objects returned by InvariantDataProvider are guaranteed to conform to the correct struct definition, so InvariantDataProvider can also be used to validate unknown data providers.

Examples

use icu_provider::prelude::*;
use icu_provider::inv::InvariantDataProvider;
use icu_provider::hello_world::{key, HelloWorldV1Marker};
use std::borrow::Cow;

let provider = InvariantDataProvider;
let result: DataPayload<HelloWorldV1Marker> = provider
    .load_payload(&DataRequest::from(key::HELLO_WORLD_V1))
    .unwrap()
    .take_payload()
    .unwrap();

assert_eq!("(und) Hello World", result.get().message);