Trait icu_provider::iter::DataExporter

pub trait DataExporter {
    fn put_payload(
        &mut self, 
        req: &DataRequest, 
        payload: &dyn ErasedDataStruct
    ) -> Result<(), Box<dyn Error>>;

    fn include_resource_options(&self, resc_options: &ResourceOptions) -> bool;

    fn put_key_from_provider<'a, 'd>(
        &mut self, 
        resc_key: &ResourceKey, 
        provider: &impl IterableErasedDataProvider<'d>
    ) -> Result<(), Error> { ... }
}

An object capable of storing/persisting data payloads to be read by a DataProvider.

A DataProvider by itself is "read-only"; this trait enables it to be "read-write".

Required methods

fn put_payload(
&mut self,
req: &DataRequest,
payload: &dyn ErasedDataStruct
) -> Result<(), Box<dyn Error>>

Save a payload corresponding to the given data request (resource path).

fn include_resource_options(&self, resc_options: &ResourceOptions) -> bool

Whether to load and dump data for the given entry. This function enables the DataExporter to filter out certain data entries.

Loading content...
Provided methods

fn put_key_from_provider<'a, 'd>(
&mut self,
resc_key: &ResourceKey,
provider: &impl IterableErasedDataProvider<'d>
) -> Result<(), Error>

Auto-implemented function that loads data from an IterableDataProvider and dumps it into this DataExporter.

Loading content...