Macro icu_provider::impl_dyn_provider

macro_rules! impl_dyn_provider {
    ($provider : ty, { $($pat : pat => $struct_m : ty), +, }, ANY) => { ... };
    ($provider : ty, [$($struct_m : ty), +,], ANY) => { ... };
    ($provider : ty, { $($pat : pat => $struct_m : ty), +, }, SERDE_SE) => { ... };
    ($provider : ty, [$($struct_m : ty), +,], SERDE_SE) => { ... };
    ($provider : ty, { $($pat : pat => $struct_m : ty), +, }, $dyn_m : path) => { ... };
    ($provider : ty, [$($struct_m : ty), +,], $dyn_m : path) => { ... };
}

Implements DynProvider for a marker type S on a type that already implements DynProvider or ResourceProvider for one or more M, where M is a concrete type that is convertible to S via UpcastDataPayload.

Use this macro to add support to your data provider for:

AnyPayload if your provider can return typed objects as Any. Use the shorthand ANY as the third argument to the macro.

SerializeMarker if your provider returns objects implementing serde::Serialize Use the shorthand SERDE_SE as the third argument to the macro.

Wrapping ResourceProvider

If your type implements ResourceProvider, pass a list of markers as the second argument. This results in a DynProvider<AnyMarker> that delegates to a specific marker if the key matches or else returns DataErrorKind::MissingResourceKey.

use icu_provider::prelude::*;
use icu_provider::iter::*;
use icu_provider::hello_world::*;
use icu_provider::inv::InvariantDataProvider;

// Example struct that implements ResourceProvider<HelloWorldV1Marker>
struct MyProvider;
impl ResourceProvider<HelloWorldV1Marker> for MyProvider {
    fn load_resource(&self, req: &DataRequest)
            -> Result<DataResponse<HelloWorldV1Marker>, DataError> {
        let provider = InvariantDataProvider;
        provider.load_resource(req)
    }
}

impl IterableResourceProvider<HelloWorldV1Marker> for MyProvider {
    fn supported_options(&self)
        -> Result<Box<dyn Iterator<Item = ResourceOptions> + '_>, DataError> {
        Ok(Box::new(core::iter::once(Default::default())))
    }
}

// Implement DataProvider<AnyMarker> on this struct
icu_provider::impl_dyn_provider!(MyProvider, [
    HelloWorldV1Marker,
], ANY);

let provider = MyProvider;

// Successful result if the key matches:
assert!(matches!(
    provider.load_payload(HelloWorldV1Marker::KEY, &Default::default()),
    Ok(_)
));

// Failure if the key does not match:
let DUMMY_KEY = icu_provider::resource_key!("dummy@1");
assert!(matches!(
    provider.load_payload(DUMMY_KEY, &Default::default()),
    Err(DataError { kind: DataErrorKind::MissingResourceKey, .. })
));

Wrapping DynProvider

It is also possible to wrap a DynProvider to create another DynProvider. To do this, pass a match-like statement for keys as the second argument:

use icu_provider::prelude::*;
use icu_provider::iter::*;
use icu_provider::hello_world::*;
use icu_provider::inv::InvariantDataProvider;

// Example struct that implements DynProvider<HelloWorldV1Marker>
struct MyProvider;
impl DynProvider<HelloWorldV1Marker> for MyProvider {
    fn load_payload(&self, key: ResourceKey, req: &DataRequest)
            -> Result<DataResponse<HelloWorldV1Marker>, DataError> {
        let provider = InvariantDataProvider;
        provider.load_resource(req)
    }
}

impl IterableDynProvider<HelloWorldV1Marker> for MyProvider {
    fn supported_options_for_key(&self, _key: &ResourceKey)
        -> Result<Box<dyn Iterator<Item = ResourceOptions> + '_>, DataError> {
        Ok(Box::new(core::iter::once(Default::default())))
    }
}

// Implement DataProvider<AnyMarker> on this struct.
// Match HelloWorldV1Marker::KEY and delegate to DynProvider<HelloWorldV1Marker>.
// Send the wildcard match also to DynProvider<HelloWorldV1Marker>.
icu_provider::impl_dyn_provider!(MyProvider, {
    HelloWorldV1Marker::KEY => HelloWorldV1Marker,
    _ => HelloWorldV1Marker,
}, ANY);

let provider = MyProvider;
let provider = provider.as_any_provider();

// Successful result if the key matches:
assert!(matches!(
    provider.load_any(HelloWorldV1Marker::KEY, &Default::default()),
    Ok(_)
));

// Because of the wildcard, non-matching requests are captured:
let DUMMY_KEY = icu_provider::resource_key!("dummy@1");
assert!(matches!(
    provider.load_any(DUMMY_KEY, &Default::default()),
    Ok(_)
));