Trait icu_provider::marker::KeyedDataMarker

pub trait KeyedDataMarker: DataMarker {
    const KEY: DataKey;
}

A DataMarker with a DataKey attached.

Structs implementing this trait are normally generated with the data_struct! macro.

Implementing this trait enables this marker to be used with the main DataProvider trait. Most markers should be associated with a specific key and should therefore implement this trait.

BufferMarker and AnyMarker are examples of markers that do not implement this trait because they are not specific to a single key.

Required Associated Constants

const KEY: DataKey

The single DataKey associated with this marker.