Module icu_provider::prelude

Core selection of APIs and structures for the ICU4X data provider.

Re-exports

pub use crate::any::AnyMarker;

pub use crate::any::AnyPayload;

pub use crate::any::AnyProvider;

pub use crate::any::AnyResponse;

pub use crate::buf::BufferMarker;

pub use crate::buf::BufferProvider;

pub use crate::marker::DataMarker;

pub use crate::marker::KeyedDataMarker;

pub use crate::any::AsDowncastingAnyProvider;

pub use crate::any::AsDynamicDataProviderAnyMarkerWrap;

Macros

data_key

See DataKey.

Structs

DataError

The error type for ICU4X data provider operations.

DataKey

Used for loading data from an ICU4X data provider.

DataKeyHash

A compact hash of a DataKey. Useful for keys in maps.

DataLocale

The main locale type used by the ICU4X data provider.

DataPayload

A container for data payloads returned from a data provider.

DataRequest

The request type passed into all data provider implementations.

DataResponse

A response object containing an object as payload and metadata about it.

DataResponseMetadata

A response object containing metadata about the returned data.

Enums

DataErrorKind

A list specifying general categories of data provider error.

Traits

DataProvider

A data provider that loads data for a specific DataKey.

DynamicDataProvider

A data provider that loads data for a specific data type.