Enum icu_provider::prelude::DataError

#[non_exhaustive]pub enum DataError {
    UnsupportedCategory(ResourceCategory),
    UnsupportedResourceKey(ResourceKey),
    UnavailableResourceOptions(DataRequest),
    NeedsLanguageIdentifier(DataRequest),
    NeedsTypeInfo,
    MissingPayload,
    MismatchedType {
        actual: Option<TypeId>,
        generic: Option<TypeId>,
    },
    Serde(Error),
    Resource(Box<dyn Error>),
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UnsupportedCategory(ResourceCategory)

The data provider does not support the requested category.

UnsupportedResourceKey(ResourceKey)

The data provider supports the category, but not the key (sub-category or version).

UnavailableResourceOptions(DataRequest)

The data provider supports the key, but does not have data for the specific entry.

NeedsLanguageIdentifier(DataRequest)

The data provider supports the key, but it requires a language identifier, which was missing from the request.

NeedsTypeInfo

The operation cannot be completed without more type information. For example, data cannot be deserialized without the concrete type.

MissingPayload

The payload is missing. This error is usually unexpected.

MismatchedType

The TypeID of the payload does not match the expected TypeID.

Fields of MismatchedType

actual: Option<TypeId>

The actual TypeID of the payload, if available.

generic: Option<TypeId>

The expected TypeID derived from the generic type parameter at the call site.

Serde(Error)

An error occured during serialization or deserialization.

Resource(Box<dyn Error>)

The data provider encountered some other error when loading the resource, such as I/O.

Implementations

impl Error

pub fn new_resc_error<T>(err: T) -> Self where
T: 'static + Error,