Struct icu_provider::prelude::DataPayload

pub struct DataPayload<'d, T: ?Sized> where
    T: ToOwned,
    <T as ToOwned>::Owned: Debug,  {
    pub cow: Option<Cow<'d, T>>,
}

A wrapper around the payload returned in a DataResponse.

Example

use icu_provider::prelude::*;
use std::borrow::Cow;

let payload = DataPayload { cow: Some(Cow::Borrowed("Demo")) };

assert_eq!("Demo", payload.borrow().expect("Payload is present"));

Fields
cow: Option<Cow<'d, T>>
Implementations

impl<'d, T: ?Sized> DataPayload<'d, T> where
T: ToOwned,
<T as ToOwned>::Owned: Debug,

pub fn new() -> Self

Creates a new, empty DataPayload.

Default is not implemented because it would be misleading: does the DataPayload start empty, or does it start with the Default value of T?

pub fn borrow(&self) -> Result<&T, Error>

Borrows the underlying payload. Error if not present.

pub fn take(&mut self) -> Result<Cow<'d, T>, Error>

Takes ownership of the underlying payload. Error if not present.

impl<'d> DataPayload<'d, dyn ErasedDataStruct>

pub fn downcast<T>(self) -> Result<DataPayload<'d, T>, Error> where
T: Clone + Debug + Any,

Convert this DataPayload of an ErasedDataStruct into a DataPayload of a Sized type. Returns an error if the type is not compatible.