Struct icu_provider::prelude::ResourceKey

pub struct ResourceKey { /* fields omitted */ }

The resource key used for loading data from an ICU4X data provider.

A resource key is tightly coupled with the code that uses it to load data at runtime. Executables can be searched for ResourceKey instances to produce optimized data files. Therefore, users should not generally create ResourceKey instances; they should instead use the ones exported by a component.

Implementations

impl ResourceKey

pub fn get_path(&self) -> &str

Gets a human-readable representation of a ResourceKey.

The human-readable path string always contains at least one ‘/’, and it ends with ‘@’ followed by one or more digits. Paths do not contain characters other than ASCII, ‘_’, ‘/’, ‘=’, and ‘@’.

Useful for reading and writing data to a file system.

pub const fn get_hash(&self) -> ResourceKeyHash

Gets a machine-readable representation of a ResourceKey.

The machine-readable hash is 4 bytes and can be used as the key in a map.

The hash is a 32-bit FxHash of the path, computed as if on a little-endian platform.

pub const fn try_new(path: &'static str) -> Result<Self, DataError>

Creates a new ResourceKey from a path, returning an error if the path is invalid.

It is intended that ResourceKey objects are const-constructed. To force construction into a const context, use resource_key!(). Doing so ensures that compile-time key extraction functions as expected.

Example

use icu_provider::prelude::*;
use icu_provider::tagged;

// Const constructed (preferred):
const k1: ResourceKey = icu_provider::resource_key!("foo/bar@1");

// Runtime constructed:
let k2: ResourceKey = ResourceKey::try_new(tagged!("foo/bar@1")).unwrap();

assert_eq!(k1, k2);

impl ResourceKey

pub fn get_last_component_no_version(&self) -> &str

Gets the last path component of a ResourceKey without the version suffix.

Examples

use icu_provider::prelude::*;

let resc_key = icu_provider::hello_world::key::HELLO_WORLD_V1;
assert_eq!("helloworld", resc_key.get_last_component_no_version());

pub fn match_key(&self, key: Self) -> Result<(), DataError>

Returns Ok if this data key matches the argument, or the appropriate error.

Convenience method for data providers that support a single ResourceKey.

Examples

use icu_provider::prelude::*;

const FOO_BAR: ResourceKey = icu_provider::resource_key!("foo/bar@1");
const FOO_BAZ: ResourceKey = icu_provider::resource_key!("foo/baz@1");
const BAR_BAZ: ResourceKey = icu_provider::resource_key!("bar/baz@1");

assert!(matches!(
    FOO_BAR.match_key(FOO_BAR),
    Ok(())
));
assert!(matches!(
    FOO_BAR.match_key(FOO_BAZ),
    Err(DataError { kind: DataErrorKind::MissingResourceKey, .. })
));
assert!(matches!(
    FOO_BAR.match_key(BAR_BAZ),
    Err(DataError { kind: DataErrorKind::MissingResourceKey, .. })
));

// The error context contains the argument:
assert_eq!(
    FOO_BAR.match_key(BAR_BAZ).unwrap_err().key,
    Some(BAR_BAZ)
);