Struct icu_provider::prelude::ResourceOptions

pub struct ResourceOptions { /* private fields */ }

A variant and language identifier, used for requesting data from a data provider.

The fields in a ResourceOptions are not generally known until runtime.

Implementations

impl ResourceOptions

pub fn cmp_bytes(&self, other: &[u8]) -> Ordering

impl ResourceOptions

pub fn temp_for_region(region: Option<Region>) -> Self

TODO(#1109): Delete this function and use vertical fallback instead

pub fn is_empty(&self) -> bool

Returns whether this ResourceOptions has all empty fields (no components).

pub fn is_langid_und(&self) -> bool

Returns whether the LanguageIdentifier associated with this request is und.

Note that this only checks the language identifier; extension keywords may also be set. To check the entire ResourceOptions, use ResourceOptions::is_empty().

pub fn get_langid(&self) -> LanguageIdentifier

Gets the LanguageIdentifier for this ResourceOptions.

This may allocate memory if there are variant subtags. If you need only the language, script, and/or region subtag, use the specific getters for those subtags:

ResourceOptions::language()

ResourceOptions::script()

ResourceOptions::region()

If you have ownership over the ResourceOptions, use ResourceOptions::into_locale() and then access the id field.

Examples

use icu_provider::prelude::*;
use icu_locid::langid;

const FOO_BAR: ResourceKey = icu_provider::resource_key!("foo/bar@1");

let req_no_langid = DataRequest {
    options: ResourceOptions::default(),
    metadata: Default::default(),
};

let req_with_langid = DataRequest {
    options: langid!("ar-EG").into(),
    metadata: Default::default(),
};

assert_eq!(req_no_langid.options.get_langid(), langid!("und"));
assert_eq!(req_with_langid.options.get_langid(), langid!("ar-EG"));

pub fn into_locale(self) -> Locale

Converts this ResourceOptions into a Locale.

See also ResourceOptions::get_langid().

Examples

use icu_provider::prelude::*;
use icu_locid::{langid, language, region, Locale};

let locale: Locale = "it-IT-u-ca-coptic".parse()
    .expect("Valid BCP-47");
let options: ResourceOptions = locale.into();

assert_eq!(options.to_string(), "it-IT-u-ca-coptic");
assert_eq!(options.get_langid(), langid!("it-IT"));
assert_eq!(options.language(), language!("it"));
assert_eq!(options.script(), None);
assert_eq!(options.region(), Some(region!("IT")));

let locale = options.into_locale();
assert_eq!(locale.to_string(), "it-IT-u-ca-coptic");

pub fn language(&self) -> Language

Returns the Language for this ResourceOptions.

pub fn script(&self) -> Option<Script>

Returns the Script for this ResourceOptions.

pub fn region(&self) -> Option<Region>

Returns the Region for this ResourceOptions.

pub fn get_unicode_ext(&self, key: &Key) -> Option<Value>

Gets the value of the specified Unicode extension keyword for this ResourceOptions.

pub fn matches_unicode_ext(&self, key: &Key, value: &Value) -> bool

Returns whether this ResourceOptions contains a Unicode extension keyword with the specified key and value.

Examples

use icu_provider::prelude::*;
use icu_locid::{Locale, unicode_ext_key, unicode_ext_value};

let locale: Locale = "it-IT-u-ca-coptic".parse()
    .expect("Valid BCP-47");
let options: ResourceOptions = locale.into();

assert_eq!(
    options.get_unicode_ext(&unicode_ext_key!("hc")),
    None
);
assert_eq!(
    options.get_unicode_ext(&unicode_ext_key!("ca")),
    Some(unicode_ext_value!("coptic"))
);
assert!(options.matches_unicode_ext(
    &unicode_ext_key!("ca"),
    &unicode_ext_value!("coptic"),
));