Trait icu_provider::prelude::DataReceiver

pub trait DataReceiver<'d, 'de> {
    fn receive_deserializer(
        &mut self, 
        deserializer: &mut dyn Deserializer<'de>
    ) -> Result<(), Error>;

    fn receive_borrow(&mut self, borrowed_any: &'d dyn Any) -> Result<(), Error>;

    fn receive_box(&mut self, boxed_any: Box<dyn Any>) -> Result<(), Error>;

    fn receive_option(&mut self, option_any: &mut dyn Any) -> Result<(), Error>;

    fn as_serialize(&self) -> Option<&dyn Serialize>;
}

An object capable of accepting data from a variety of forms. Lifetimes:

'd = lifetime of borrowed data (Cow::Borrowed)

'de = lifetime parameter of owned data (Cow::Owned)

Required methods

fn receive_deserializer(
&mut self,
deserializer: &mut dyn Deserializer<'de>
) -> Result<(), Error>

Consumes a Serde Deserializer into this DataReceiver as owned data.

Example

use icu_provider::prelude::*;
use std::borrow::Cow;

const JSON: &'static str = "\"hello world\"";

let mut receiver = DataReceiverForType::<String>::new();
let mut d = serde_json::Deserializer::from_str(JSON);
receiver.receive_deserializer(&mut erased_serde::Deserializer::erase(&mut d))
    .expect("Deserialization should be successful");

assert_eq!(receiver.payload, Some(Cow::Owned("hello world".to_string())));

fn receive_borrow(&mut self, borrowed_any: &'d dyn Any) -> Result<(), Error>

Sets this DataReceiver to a borrowed value.

Example

use icu_provider::prelude::*;
use std::borrow::Cow;

let borrowed_data: String = "hello world".to_string();

let mut receiver = DataReceiverForType::<String>::new();
receiver.receive_borrow(&borrowed_data).expect("Types should match");

assert_eq!(receiver.payload, Some(Cow::Borrowed(&borrowed_data)));

fn receive_box(&mut self, boxed_any: Box<dyn Any>) -> Result<(), Error>

Takes a Box into this DataReceiver as owned data.

Example

use icu_provider::prelude::*;
use std::borrow::Cow;

let mut receiver = DataReceiverForType::<String>::new();
let boxed = Box::new("hello world".to_string());
receiver.receive_box(boxed).expect("Types should match");

assert_eq!(receiver.payload, Some(Cow::Owned("hello world".to_string())));

fn receive_option(&mut self, option_any: &mut dyn Any) -> Result<(), Error>

Takes the value out of an Option into this DataReceiver as owned data.

Example

use icu_provider::prelude::*;
use std::borrow::Cow;

let mut receiver = DataReceiverForType::<String>::new();
let mut option = Some("hello world".to_string());
receiver.receive_option(&mut option).expect("Types should match");

assert!(option.is_none());
assert_eq!(receiver.payload, Some(Cow::Owned("hello world".to_string())));

fn as_serialize(&self) -> Option<&dyn Serialize>

Borrows the value in this DataReceiver as a Serialize trait object.

Loading content...
Implementations

impl<'a, 'd, 'de> dyn DataReceiver<'d, 'de> + 'a

pub fn receive_invariant<T: Default + Any>(&mut self) -> Result<(), Error>

Sets this DataReceiver to the default value of the given type.