Trait icu_provider::prelude::ResourceProvider

pub trait ResourceProvider<M> where
    M: ResourceMarker, {
    fn load_resource(
        &self, 
        req: &DataRequest
    ) -> Result<DataResponse<M>, DataError>;
}

A data provider that loads data for a specific ResourceKey.

Required methods

fn load_resource(&self, req: &DataRequest) -> Result<DataResponse<M>, DataError>

Query the provider for data, returning the result.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.