icu_provider

Module request

Structs

AttributeParseError

Invalid character

DataIdentifierBorrowed

The borrowed version of a DataIdentifierCow.

DataIdentifierCow

A data identifier identifies a particular version of data, such as “English”.

DataLocale

A locale type optimized for use in fallbacking and the ICU4X data pipeline.

DataMarkerAttributes

An additional key to identify data beyond a DataLocale.

DataRequest

The request type passed into all data provider implementations.

DataRequestMetadata

Metadata for data requests. This is currently empty, but it may be extended with options for tuning locale fallback, buffer layout, and so forth.