Module icu_provider::resource

Macros

tinystr4
Re-export tinystr4 for crate macro resource_key!()

tinystr16
Re-export tinystr16 for crate macro resource_key!()

Structs

ResourceKey
A category, subcategory, and version, used for requesting data from a DataProvider.

ResourceKeyComponents
The standard components of a ResourceKey path.

ResourceOptions
A variant and language identifier, used for requesting data from a DataProvider.

ResourceOptionsComponents
The standard components of a ResourceOptions path.

ResourcePath

Enums

ResourceCategory
A top-level collection of related resource keys.