Trait icu_provider::serde::SerdeDeDataReceiver

pub trait SerdeDeDataReceiver<'de> {
    fn receive_deserializer(
        &mut self, 
        deserializer: &mut dyn Deserializer<'de>
    ) -> Result<(), Error>;
}

An object that receives data from a Serde Deserializer. Implemented by DataPayload.

Lifetimes:

'de = deserializer lifetime; can usually be '_

Required methods

fn receive_deserializer(
&mut self,
deserializer: &mut dyn Deserializer<'de>
) -> Result<(), Error>

Consumes a Serde Deserializer into this SerdeDeDataReceiver as owned data.

This method results in an owned payload, but the payload could have non-static references according to the deserializer lifetime.

Examples

use icu_provider::prelude::*;
use icu_provider::serde::SerdeDeDataReceiver;

const JSON: &'static str = "\"hello world\"";

let mut receiver: Option<&str> = None;
let mut d = serde_json::Deserializer::from_str(JSON);
receiver.receive_deserializer(&mut erased_serde::Deserializer::erase(&mut d))
    .expect("Deserialization should be successful");

assert!(matches!(receiver, Some(_)));
assert_eq!(receiver, Some("hello world"));

Loading content...