Struct icu_provider::AuxiliaryKeys

pub struct AuxiliaryKeys { /* private fields */ }

The โauxiliary keyโ is an annotation on DataLocale that can contain an arbitrary information that does not fit into the LanguageIdentifier or Keywords.

A DataLocale can have multiple auxiliary keys, represented by this struct. The auxiliary keys are stored as private use subtags following -x-.

An auxiliary key currently allows 1-8 lowercase alphanumerics.

๐ง This code is experimental; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. It can be enabled with the "experimental" Cargo feature of the `icu_provider` crate. Use with caution. #3632

Examples

use icu_locid::locale;
use icu_provider::prelude::*;
use writeable::assert_writeable_eq;

let mut data_locale: DataLocale = locale!("ar-EG").into();
assert_writeable_eq!(data_locale, "ar-EG");
assert!(!data_locale.has_aux());
assert_eq!(data_locale.get_aux(), None);

let aux = "gbp"
    .parse::<AuxiliaryKeys>()
    .expect("contains valid characters");

data_locale.set_aux(aux);
assert_writeable_eq!(data_locale, "ar-EG-x-gbp");
assert!(data_locale.has_aux());
assert_eq!(data_locale.get_aux(), Some(&"gbp".parse().unwrap()));

Multiple auxiliary keys are allowed:

use icu_locid::locale;
use icu_provider::prelude::*;
use writeable::assert_writeable_eq;

let data_locale = "ar-EG-x-gbp-long".parse::<DataLocale>().unwrap();
assert_writeable_eq!(data_locale, "ar-EG-x-gbp-long");
assert_eq!(data_locale.get_aux().unwrap().iter().count(), 2);

Not all strings are valid auxiliary keys. The string must be well-formed and case-normalized:

use icu_provider::prelude::*;

assert!("abcdefg".parse::<AuxiliaryKeys>().is_ok());
assert!("abc-xyz".parse::<AuxiliaryKeys>().is_ok());

assert!("".parse::<AuxiliaryKeys>().is_err());
assert!("!@#$%".parse::<AuxiliaryKeys>().is_err());
assert!("abc_xyz".parse::<AuxiliaryKeys>().is_err());
assert!("ABC123".parse::<AuxiliaryKeys>().is_err());

Implementations

impl AuxiliaryKeys

pub fn try_from_iter( iter: impl IntoIterator<Item = Subtag> ) -> Result<Self, DataError>

Creates an AuxiliaryKeys from an iterator of individual keys.

Examples

use icu_locid::extensions::private::subtag;
use icu_provider::prelude::*;

// Single auxiliary key:
let a = AuxiliaryKeys::try_from_iter([subtag!("abc")]).unwrap();
let b = "abc".parse::<AuxiliaryKeys>().unwrap();
assert_eq!(a, b);

// Multiple auxiliary keys:
let a = AuxiliaryKeys::try_from_iter([subtag!("abc"), subtag!("defg")])
    .unwrap();
let b = "abc-defg".parse::<AuxiliaryKeys>().unwrap();
assert_eq!(a, b);

The iterator canโt be empty:

use icu_provider::prelude::*;

assert!(AuxiliaryKeys::try_from_iter([]).is_err());

pub const fn from_subtag(input: Subtag) -> Self

Creates an AuxiliaryKeys from a single subtag.

Examples

use icu_locid::extensions::private::subtag;
use icu_provider::prelude::*;

// Single auxiliary key:
let a = AuxiliaryKeys::from_subtag(subtag!("abc"));
let b = "abc".parse::<AuxiliaryKeys>().unwrap();
assert_eq!(a, b);

pub fn iter(&self) -> impl Iterator<Item = Subtag> + '_

Iterates over the components of the auxiliary key.

Example

use icu_locid::extensions::private::subtag;
use icu_provider::AuxiliaryKeys;

let aux: AuxiliaryKeys = "abc-defg".parse().unwrap();
assert_eq!(
    aux.iter().collect::<Vec<_>>(),
    vec![subtag!("abc"), subtag!("defg")]
);