icu_provider

Struct DataIdentifierCow

#[non_exhaustive]
pub struct DataIdentifierCow<'a> {
    pub marker_attributes: Cow<'a, DataMarkerAttributes>,
    pub locale: DataLocale,
}

A data identifier identifies a particular version of data, such as “English”.

It is a wrapper around a DataLocale and a DataMarkerAttributes.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

marker_attributes: Cow<'a, DataMarkerAttributes>

Marker-specific request attributes

locale: DataLocale

The CLDR locale

Implementations

impl<'a> DataIdentifierCow<'a>

pub fn as_borrowed(&'a self) -> DataIdentifierBorrowed<'a>

Borrows this DataIdentifierCow as a DataIdentifierBorrowed<'a>.

pub fn from_locale(locale: DataLocale) -> Self

Creates a DataIdentifierCow from an owned DataLocale.

pub fn from_marker_attributes( marker_attributes: &'a DataMarkerAttributes, ) -> Self

Creates a DataIdentifierCow from a borrowed DataMarkerAttributes.

pub fn from_marker_attributes_owned( marker_attributes: Box<DataMarkerAttributes>, ) -> Self

Creates a DataIdentifierCow from an owned DataMarkerAttributes.

pub fn from_owned( marker_attributes: Box<DataMarkerAttributes>, locale: DataLocale, ) -> Self

Creates a DataIdentifierCow from an owned DataMarkerAttributes and an owned DataLocale.

pub fn from_borrowed_and_owned( marker_attributes: &'a DataMarkerAttributes, locale: DataLocale, ) -> Self

Creates a DataIdentifierCow from a borrowed DataMarkerAttributes and an owned DataLocale.

pub fn is_unknown(&self) -> bool

Returns whether this id is equal to the default.