Struct icu_provider::DataKey

pub struct DataKey { /* private fields */ }

Used for loading data from an ICU4X data provider.

A resource key is tightly coupled with the code that uses it to load data at runtime. Executables can be searched for DataKey instances to produce optimized data files. Therefore, users should not generally create DataKey instances; they should instead use the ones exported by a component.

DataKeys are created with the data_key! macro:

const K: DataKey = icu_provider::data_key!("foo/bar@1");

The human-readable path string ends with @ followed by one or more digits (the version number). Paths do not contain characters other than ASCII letters and digits, _, /.

Invalid paths are compile-time errors (as data_key! uses const).

const K: DataKey = icu_provider::data_key!("foo/../bar@1");

Implementations

impl DataKey

pub const fn path(self) -> DataKeyPath

Gets a human-readable representation of a DataKey.

The human-readable path string ends with @ followed by one or more digits (the version number). Paths do not contain characters other than ASCII letters and digits, _, /.

Useful for reading and writing data to a file system.

pub const fn hashed(self) -> DataKeyHash

Gets a platform-independent hash of a DataKey.

The hash is 4 bytes and allows for fast key comparison.

Example

use icu_provider::DataKey;
use icu_provider::DataKeyHash;

const KEY: DataKey = icu_provider::data_key!("foo@1");
const KEY_HASH: DataKeyHash = KEY.hashed();

assert_eq!(KEY_HASH.to_bytes(), [0xe2, 0xb6, 0x17, 0x71]);

pub const fn metadata(self) -> DataKeyMetadata

Gets the metadata associated with this DataKey.

pub const fn fallback_config(self) -> LocaleFallbackConfig

Returns the [LocaleFallbackConfig] for this DataKey.

pub const fn from_path_and_metadata( path: DataKeyPath, metadata: DataKeyMetadata, ) -> Self

Constructs a DataKey from a path and metadata.

Examples

use icu_provider::data_key;
use icu_provider::DataKey;

const CONST_KEY: DataKey = data_key!("foo@1");

let runtime_key =
    DataKey::from_path_and_metadata(CONST_KEY.path(), CONST_KEY.metadata());

assert_eq!(CONST_KEY, runtime_key);

pub fn match_key(self, key: Self) -> Result<(), DataError>

Returns Ok if this data key matches the argument, or the appropriate error.

Convenience method for data providers that support a single DataKey.

Examples

use icu_provider::prelude::*;

const FOO_BAR: DataKey = icu_provider::data_key!("foo/bar@1");
const FOO_BAZ: DataKey = icu_provider::data_key!("foo/baz@1");
const BAR_BAZ: DataKey = icu_provider::data_key!("bar/baz@1");

assert!(matches!(FOO_BAR.match_key(FOO_BAR), Ok(())));
assert!(matches!(
    FOO_BAR.match_key(FOO_BAZ),
    Err(DataError {
        kind: DataErrorKind::MissingDataKey,
        ..
    })
));
assert!(matches!(
    FOO_BAR.match_key(BAR_BAZ),
    Err(DataError {
        kind: DataErrorKind::MissingDataKey,
        ..
    })
));

// The error context contains the argument:
assert_eq!(FOO_BAR.match_key(BAR_BAZ).unwrap_err().key, Some(BAR_BAZ));