icu_provider

Trait DataMarker

pub trait DataMarker: DynamicDataMarker {
    const INFO: DataMarkerInfo;
}

A DynamicDataMarker with a DataMarkerInfo attached.

Structs implementing this trait are normally generated with the data_struct! macro.

Implementing this trait enables this marker to be used with the main DataProvider trait. Most markers should be associated with a specific marker and should therefore implement this trait.

BufferMarker is an example of a marker that does not implement this trait.

Note: DataMarkers are quasi-const-generic compile-time objects, and as such are expected to be unit structs. As this is not something that can be enforced by the type system, we currently only have a 'static bound on them (which is needed by a lot of our code).

Required Associated Constants

const INFO: DataMarkerInfo

The single DataMarkerInfo associated with this marker.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.