icu_provider
Trait DynamicDataProvider

pub trait DynamicDataProvider<M>
where
    M: DynamicDataMarker,
{
    // Required method
    fn load_data(
        &self,
        marker: DataMarkerInfo,
        req: DataRequest<'_>,
    ) -> Result<DataResponse<M>, DataError>;
}

A data provider that loads data for a specific data type.

Unlike DataProvider, there may be multiple markers corresponding to the same data type.

Required Methods

fn load_data( &self, marker: DataMarkerInfo, req: DataRequest<'_>, ) -> Result<DataResponse<M>, DataError>

Query the provider for data, returning the result.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.