Crate icu_provider

icu_provider is one of the ICU4X components.

It defines traits and structs for transmitting data through the ICU4X locale data pipeline. The primary trait is DataProvider. It has one method, which transforms a Request into a Response:

fn load_payload(&self, req: &DataRequest) -> Result<DataResponse<'d>, DataError>

A Request contains a ResourceKey (a composition of a Category and sub-category, e.g., "plurals/cardinal@1") and ResourceOptions (a language identifier and optional variant, e.g., "fr") being requested. The Response contains the data payload corresponding to the Request.

The most common types required for ICU4X DataProvider are included via the prelude:

use icu_provider::prelude::*;

Concrete Implementations of Data Providers

Any object implementing DataProvider can be used to supply ICU4X with locale data. ICU4X ships with some pre-built data providers:

CldrJsonDataProvider reads structured data directly from CLDR source files.

FsDataProvider reads structured data from the filesystem. It can also write out that filesystem structure. More efficient than CldrJsonDataProvider.

This crate also contains some concrete implementations for testing purposes:

InvariantDataProvider returns fixed data that does not vary by locale.

StructProvider wraps a particular instance of a struct and returns it.

HelloWorldProvider returns "hello world" strings in several languages.

Additional Traits

IterableDataProvider

Data providers can implement IterableDataProvider, allowing iteration over all ResourceOptions instances supported for a certain key in the data provider.

For more information, see the iter module.

ErasedDataProvider

The DataProvider trait has a type argument corresponding to the type being loaded. A peer trait ErasedDataProvider removes the type argument, using runtime type checking instead.

Since ErasedDataProvider is not specific to a single type, it can be useful for:

Caches

Bulk data operations

Transforming from one format to another

Types and Lifetimes

All types T compatible with Cow and Debug can be passed through the data provider.

Most DataProvider traits take a lifetime argument 'd. This represents the lifetime of data returned by the DataProvider.

DataProvider returns data in the form of a Cow<'d, T>, where 'd is the lifetime of the data if it is borrowed, and T is constrained by T: 'd such that if the data is owned, it may not have any fields whose lifetime subceeds 'd.

When using ErasedDataProvider, the following additional requirements are placed on T:

T: 'static, since T must be compatible with Any

serde::Deserialize and serde::Serialize, allowing for type-agnostic (de)serialization

Default, allowing InvariantDataProvider to work

Re-exports

pub use prelude::*;

Modules

data_provider

erased

hello_world

inv

iter
Collection of iteration APIs for DataProvider.

prelude
Core selection of APIs and structures for DataProvider.

resource

struct_provider

structs
Collection of data structures for DataProvider.

Macros

impl_erased
Helper macro to implement ErasedDataProvider on an object implementing DataProvider for a single type. Calls to self.load_to_receiver delegate to self.load_payload.

resource_key
Shortcut to construct a const resource identifier.

Enums

DataError