Trait icu_provider::erased::ErasedDataProvider

pub trait ErasedDataProvider {
    fn load_erased(
        &self, 
        req: &DataRequest
    ) -> Result<DataResponse<ErasedDataStructMarker>, Error>;
}

A type-erased data provider that loads a payload of types implementing Any.

Note: This trait is redundant with DataProvider<dyn ErasedDataStruct> and auto-implemented for all types implementing that trait. This trait may eventually be removed when the following Rust issues are resolved:

#41517 (trait aliases are not supported)

#68636 (identical traits can’t be auto-implemented)

Required methods

fn load_erased(
&self,
req: &DataRequest
) -> Result<DataResponse<ErasedDataStructMarker>, Error>

Query the provider for data, returning the result as an ErasedDataStruct trait object.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.