Trait icu_provider::export::DataExporter

pub trait DataExporter<M>: Sync where
    M: DataMarker, {
    fn put_payload(
        &self, 
        key: ResourceKey, 
        options: ResourceOptions, 
        payload: DataPayload<M>
    ) -> Result<(), DataError>;

    fn flush(&self, _key: ResourceKey) -> Result<(), DataError> { ... }
    fn close(&mut self) -> Result<(), DataError> { ... }
}

An object capable of exporting data payloads in some form.

Required methods

fn put_payload(
&self,
key: ResourceKey,
options: ResourceOptions,
payload: DataPayload<M>
) -> Result<(), DataError>

Save a payload corresponding to the given key and options. Takes non-mut self as it can be called concurrently.

Provided methods

fn flush(&self, _key: ResourceKey) -> Result<(), DataError>

Function called after all keys have been fully dumped. Takes non-mut self as it can be called concurrently.

fn close(&mut self) -> Result<(), DataError>

This function has to be called before the object is dropped (after all keys have been fully dumped). This conceptually takes ownership, so clients may not interact with this object after close has been called.