Trait icu_provider::iter::IterableDataProviderCore

pub trait IterableDataProviderCore {
    fn supported_options_for_key(
        &self, 
        resc_key: &ResourceKey
    ) -> Result<Box<dyn Iterator<Item = ResourceOptions>>, Error>;
}

A provider that can iterate over all supported ResourceOptions for a certain key.

Implementing this trait means that a DataProvider knows all of the data it can successfully return from a load request.

Required methods

fn supported_options_for_key(
&self,
resc_key: &ResourceKey
) -> Result<Box<dyn Iterator<Item = ResourceOptions>>, Error>

Given a ResourceKey, returns a boxed iterator over ResourceOptions.