Macro icu_provider::impl_dyn_provider

macro_rules! impl_dyn_provider {
    ($provider:ty, { $($pat:pat => $struct_m:ty),+, }, ANY) => { ... };
    ($provider:ty, { $($pat:pat => $struct_m:ty),+, }, SERDE_SE) => { ... };
    ($provider:ty, { $($pat:pat => $struct_m:ty),+, }, $dyn_m:path) => { ... };
}

Implements DataProvider for a marker type S on a type that already implements DataProvider for one or more M, where M is a concrete type that is convertible to S via UpcastDataPayload.

Use this macro to add support to your data provider for:

AnyPayload if your provider can return typed objects as Any

SerializeMarker if your provider returns objects implementing serde::Serialize

The second argument is a match-like construction mapping from resource keys to structs. To map multiple keys to a single data struct, use _ as the data key.

The third argument can be either the trait object marker, like SerializeMarker, or the shorthands ANY or SERDE_SE.

Lifetimes:

$data is the lifetime parameter for DataProvider; usually 'data.

Examples

Basic usage:

use icu_provider::prelude::*;
use icu_provider::marker::CowStrMarker;
use std::borrow::Cow;
const DEMO_KEY: ResourceKey = icu_provider::resource_key!("foo/bar@1");

// A small DataProvider that returns owned strings
struct MyProvider(pub String);
impl DataProvider<CowStrMarker> for MyProvider {
    fn load_payload(&self, req: &DataRequest)
            -> Result<DataResponse<CowStrMarker>, DataError> {
        req.resource_path.key.match_key(DEMO_KEY)?;
        Ok(DataResponse {
            metadata: Default::default(),
            payload: Some(DataPayload::from_owned(Cow::Owned(self.0.to_string())))
        })
    }
}

// Implement DataProvider<AnyMarker>
icu_provider::impl_dyn_provider!(MyProvider, {
    DEMO_KEY => CowStrMarker,
}, ANY);

// Usage example
let provider = MyProvider("demo".to_string());
let resp: DataResponse<AnyMarker> = provider
    .load_payload(&DEMO_KEY.into())
    .expect("Loading should succeed");
let payload: DataPayload<CowStrMarker> = resp
    .take_payload()
    .expect("Payload should be present")
    .downcast()
    .expect("Type should downcast successfully");
assert_eq!("demo", payload.get());

Using the wildcard _ match:

// Send all keys to the `CowStrMarker` provider.
icu_provider::impl_dyn_provider!(MyProvider, {
    _ => CowStrMarker,
}, ANY);