Module icu_provider::prelude

Core selection of APIs and structures for the ICU4X data provider.

Re-exports

pub use crate::any::AnyMarker;

pub use crate::any::AnyPayload;

pub use crate::any::AnyProvider;

pub use crate::any::AnyResponse;

pub use crate::buf::BufferMarker;

pub use crate::buf::BufferProvider;

pub use crate::marker::DataMarker;

pub use crate::marker::ResourceMarker;

pub use crate::resource_key;

pub use crate::any::AsDowncastingAnyProvider;

pub use crate::any::AsDynProviderAnyMarkerWrap;

pub use yoke;

pub use zerofrom;

Structs

DataError

The error type for ICU4X data provider operations.

DataPayload

A container for data payloads returned from a data provider.

DataRequest

DataResponse

A response object containing an object as payload and metadata about it.

DataResponseMetadata

A response object containing metadata about the returned data.

ResourceKey

Used for loading data from an ICU4X data provider.

ResourceKeyHash

A compact hash of a ResourceKey. Useful for keys in maps.

ResourceOptions

A variant and language identifier, used for requesting data from a data provider.

Enums

DataErrorKind

A list specifying general categories of data provider error.

Traits

AsDeserializingBufferProvider

Auto-implemented for all BufferProvider for easy wrapping in DeserializingBufferProvider.

DynProvider

A data provider that loads data for a specific data type.

ResourceProvider

A data provider that loads data for a specific ResourceKey.