Module icu_provider::prelude

Core selection of APIs and structures for DataProvider.

Re-exports

pub use crate::marker::DataMarker;

Structs

DataPayload
A wrapper around the payload returned in a DataResponse.

DataRequest
A struct to request a certain piece of data from a data provider.

DataResponse
A response object containing an object as payload and metadata about it.

DataResponseMetadata
A response object containing metadata about the returned data.

ResourceKey
A category, subcategory, and version, used for requesting data from a DataProvider.

ResourceOptions
A variant and language identifier, used for requesting data from a DataProvider.

ResourcePath

Enums

DataError
Error enumeration for DataProvider.

ResourceCategory
A top-level collection of related resource keys.

Traits

DataProvider
A generic data provider that loads a payload of a specific type.