Struct icu_provider::prelude::DataPayload

pub struct DataPayload<'data, M> where
    M: DataMarker<'data>,  { /* fields omitted */ }

A wrapper around the payload returned in a DataResponse.

Internally, the data is represented using the yoke crate, with several variations for different ownership models.

DataPayload is closely coupled with DataMarker.

Examples

Basic usage, using the CowStrMarker marker:

use icu_provider::prelude::*;
use icu_provider::marker::CowStrMarker;
use std::borrow::Cow;

let payload = DataPayload::<CowStrMarker>::from_owned(Cow::Borrowed("Demo"));

assert_eq!("Demo", payload.get());

Implementations

impl<'data, M> DataPayload<'data, M> where
M: DataMarker<'data>,
M::Yokeable: ZeroCopyFrom<M::Cart>,

pub fn from_partial_owned(data: Rc<M::Cart>) -> Self

Convert an Rc<Cart> into a DataPayload.

The data need not be fully owned; this constructor creates payloads bounded by 'data.

Examples

use icu_provider::prelude::*;
use icu_provider::hello_world::*;
use std::borrow::Cow;
use std::rc::Rc;

let local_data = "example".to_string();

let rc_struct = Rc::from(HelloWorldV1 {
    message: Cow::Borrowed(&local_data),
});

let payload = DataPayload::<HelloWorldV1Marker>::from_partial_owned(rc_struct.clone());

assert_eq!(payload.get(), &*rc_struct);

impl<'data, M> DataPayload<'data, M> where
M: DataMarker<'data>,

pub fn try_from_rc_buffer<E>(
rc_buffer: Rc<[u8]>,
f: impl for<'de> FnOnce(&'de [u8]) -> Result<<M::Yokeable as Yokeable<'de>>::Output, E>
) -> Result<Self, E>

Convert a byte buffer into a DataPayload. A function must be provided to perform the conversion. This can often be a Serde deserialization operation.

This constructor creates 'static payloads; borrowing is handled by Yoke.

Due to compiler bug #84937, call sites for this function may not compile; if this happens, use try_from_rc_buffer_badly() instead.

pub fn try_from_rc_buffer_badly<E>(
rc_buffer: Rc<[u8]>,
f: for<'de> fn(_: &'de [u8]) -> Result<<M::Yokeable as Yokeable<'de>>::Output, E>
) -> Result<Self, E>

Convert a byte buffer into a DataPayload. A function must be provided to perform the conversion. This can often be a Serde deserialization operation.

This constructor creates 'static payloads; borrowing is handled by Yoke.

For a version of this function that takes a FnOnce instead of a raw function pointer, see try_from_rc_buffer().

Examples

use icu_provider::prelude::*;
use icu_provider::hello_world::*;
use std::rc::Rc;
use icu_provider::yoke::Yokeable;

let json_text = "{\"message\":\"Hello World\"}";
let json_rc_buffer: Rc<[u8]> = json_text.as_bytes().into();

let payload = DataPayload::<HelloWorldV1Marker>::try_from_rc_buffer_badly(
    json_rc_buffer.clone(),
    |bytes| {
        serde_json::from_slice(bytes)
    }
)
.expect("JSON is valid");

assert_eq!("Hello World", payload.get().message);

pub fn from_owned(data: M::Yokeable) -> Self

Convert a fully owned ('static) data struct into a DataPayload.

This constructor creates 'static payloads.

Examples

use icu_provider::prelude::*;
use icu_provider::hello_world::*;
use std::borrow::Cow;

let local_struct = HelloWorldV1 {
    message: Cow::Owned("example".to_string()),
};

let payload = DataPayload::<HelloWorldV1Marker>::from_owned(local_struct.clone());

assert_eq!(payload.get(), &local_struct);

pub fn with_mut<'a, F>(&'a mut self, f: F) where
F: 'static + for<'b> FnOnce(&'b mut <M::Yokeable as Yokeable<'a>>::Output),

Mutate the data contained in this DataPayload.

For safety, all mutation operations must take place within a helper function that cannot borrow data from the surrounding context.

Examples

Basic usage:

use icu_provider::prelude::*;
use icu_provider::marker::CowStrMarker;

let mut payload = DataPayload::<CowStrMarker>::from_static_str("Hello");

payload.with_mut(|s| s.to_mut().push_str(" World"));

assert_eq!("Hello World", payload.get());

To transfer data from the context into the data struct, use the move keyword:

use icu_provider::prelude::*;
use icu_provider::marker::CowStrMarker;

let mut payload = DataPayload::<CowStrMarker>::from_static_str("Hello");

let suffix = " World".to_string();
payload.with_mut(move |s| s.to_mut().push_str(&suffix));

assert_eq!("Hello World", payload.get());

pub fn get<'a>(&'a self) -> &'a <M::Yokeable as Yokeable<'a>>::Output

Borrows the underlying data.

This function should be used like Deref would normally be used. For more information on why DataPayload cannot implement Deref, see the yoke crate.

Examples

use icu_provider::prelude::*;
use icu_provider::marker::CowStrMarker;

let payload = DataPayload::<CowStrMarker>::from_static_str("Demo");

assert_eq!("Demo", payload.get());

impl<'data> DataPayload<'static, ErasedDataStructMarker>

pub fn downcast<M>(self) -> Result<DataPayload<'static, M>, Error> where
M: DataMarker<'static>,
M::Cart: Sized,
M::Yokeable: ZeroCopyFrom<M::Cart>,

Convert this DataPayload of an ErasedDataStruct into a DataPayload of a concrete type.

Returns an error if the type is not compatible.

This is the main way to consume data returned from an ErasedDataProvider.

Internally, this method reverses the transformation performed by UpcastDataPayload::upcast as implemented for ErasedDataStructMarker.

Examples

use icu_provider::prelude::*;
use icu_provider::erased::*;
use icu_provider::hello_world::*;
use icu_locid_macros::langid;

let provider = HelloWorldProvider::new_with_placeholder_data();

let erased_payload: DataPayload<ErasedDataStructMarker> = provider
    .load_payload(&DataRequest {
        resource_path: ResourcePath {
            key: key::HELLO_WORLD_V1,
            options: ResourceOptions {
                variant: None,
                langid: Some(langid!("de")),
            }
        }
    })
    .expect("Loading should succeed")
    .take_payload()
    .expect("Data should be present");

let downcast_payload: DataPayload<HelloWorldV1Marker> = erased_payload
    .downcast()
    .expect("Types should match");

assert_eq!("Hallo Welt", downcast_payload.get().message);

impl DataPayload<'static, CowStrMarker>

pub fn from_static_str(s: &'static str) -> DataPayload<'static, CowStrMarker>

Make a DataPayload<CowStrMarker> from a static string slice.