Trait icu_provider::prelude::DataProvider

pub trait DataProvider<'data, M> where
    M: DataMarker<'data>, {
    fn load_payload(
        &self, 
        req: &DataRequest
    ) -> Result<DataResponse<'data, M>, Error>;
}

A generic data provider that loads a payload of a specific type.

See examples on some of the concrete implementations:

HelloWorldProvider

StructProvider

InvariantDataProvider

Required methods

fn load_payload(
&self,
req: &DataRequest
) -> Result<DataResponse<'data, M>, Error>

Query the provider for data, returning the result.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.

Loading content...