Trait icu_provider::serde::SerdeDeDataProvider

pub trait SerdeDeDataProvider<'de> {
    fn load_to_receiver(
        &self, 
        req: &DataRequest, 
        receiver: &mut dyn SerdeDeDataReceiver<'de>
    ) -> Result<DataResponseMetadata, Error>;
}

A type-erased data provider that loads payloads from a Serde Deserializer.

Uses erased_serde to allow the trait to be object-safe.

Required methods

fn load_to_receiver(
&self,
req: &DataRequest,
receiver: &mut dyn SerdeDeDataReceiver<'de>
) -> Result<DataResponseMetadata, Error>

Query the provider for data, loading it into a SerdeDeDataReceiver.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.

Loading content...