Trait icu_provider::serde::SerdeDeDataReceiver

pub trait SerdeDeDataReceiver {
    fn receive_rc_buffer(
        &mut self, 
        rc_buffer: Rc<[u8]>, 
        f1: for<'de> fn(bytes: &'de [u8], f2: &mut dyn FnMut(&mut dyn Deserializer<'de>))
    ) -> Result<(), Error>;

    fn receive_static(
        &mut self, 
        deserializer: &mut dyn Deserializer<'static>
    ) -> Result<(), Error>;
}

An object that receives data from a Serde Deserializer.

Implemented by Option<DataPayload>.

Required methods

fn receive_rc_buffer(
&mut self,
rc_buffer: Rc<[u8]>,
f1: for<'de> fn(bytes: &'de [u8], f2: &mut dyn FnMut(&mut dyn Deserializer<'de>))
) -> Result<(), Error>

Receives a reference-counted byte buffer.

Upon calling this function, the receiver sends byte buffer back to the caller as the first argument of f1. The caller should then map the byte buffer to an erased_serde::Deserializer and pass it back to the receiver via f2.

Examples

Deserialize from a reference-counted buffer:

use icu_provider::prelude::*;
use icu_provider::hello_world::*;
use icu_provider::serde::SerdeDeDataReceiver;
use std::rc::Rc;

let json_text = "{\"message\":\"Hello World\"}";
let rc_buffer: Rc<[u8]> = json_text.as_bytes().into();
let mut receiver: Option<DataPayload<HelloWorldV1Marker>> = None;
receiver
    .receive_rc_buffer(rc_buffer, |bytes, f2| {
        let mut d = serde_json::Deserializer::from_slice(bytes);
        f2(&mut erased_serde::Deserializer::erase(&mut d))
    })
    .expect("Well-formed data");
let payload = receiver.expect("Data is present");

assert_eq!(payload.get().message, "Hello World");

fn receive_static(
&mut self,
deserializer: &mut dyn Deserializer<'static>
) -> Result<(), Error>

Receives a &'static byte buffer via an erased_serde::Deserializer.

Note: Since the purpose of this function is to handle zero-copy deserialization of static byte buffers, we want Deserializer<'static> as opposed to DeserializeOwned.

Examples

Deserialize from a string to create static references:

use icu_provider::prelude::*;
use icu_provider::hello_world::*;
use icu_provider::serde::SerdeDeDataReceiver;

let json_text = "{\"message\":\"Hello World\"}";
let deserializer = &mut serde_json::Deserializer::from_str(json_text);
let mut receiver: Option<DataPayload<HelloWorldV1Marker>> = None;
receiver
    .receive_static(&mut erased_serde::Deserializer::erase(deserializer))
    .expect("Well-formed data");
let payload = receiver.expect("Data is present");

assert_eq!(payload.get().message, "Hello World");

Loading content...