icu_provider

Struct Cart

pub struct Cart(/* private fields */);

The type of the โcartโ that is used by DataPayload.

This type is public but the inner cart type is private. To create a Yoke with this cart, use Cart::try_make_yoke. Then, convert it to a DataPayload with DataPayload::from_yoked_buffer.

Implementations

impl Cart

pub fn try_make_yoke<Y, F, E>( cart: Box<[u8]>, f: F, ) -> Result<Yoke<Y, Option<Self>>, E>
where for<'a> Y: Yokeable<'a>, F: FnOnce(&[u8]) -> Result<<Y as Yokeable<'_>>::Output, E>,

Creates a Yoke<Y, Option<Cart>> from owned bytes by applying f.