icu_provider

Trait DataProvider

pub trait DataProvider<M>
where
    M: KeyedDataMarker,
{
    // Required method
    fn load(&self, req: DataRequest<'_>) -> Result<DataResponse<M>, DataError>;
}

A data provider that loads data for a specific DataKey.

Required Methods

fn load(&self, req: DataRequest<'_>) -> Result<DataResponse<M>, DataError>

Query the provider for data, returning the result.

Returns Ok if the request successfully loaded data. If data failed to load, returns an Error with more information.