icu_provider

Trait DryDataProvider

pub trait DryDataProvider<M: DataMarker>: DataProvider<M> {
    // Required method
    fn dry_load(
        &self,
        req: DataRequest<'_>,
    ) -> Result<DataResponseMetadata, DataError>;
}

A data provider that can determine whether it can load a particular data identifier, potentially cheaper than actually performing the load.

Required Methods

fn dry_load( &self, req: DataRequest<'_>, ) -> Result<DataResponseMetadata, DataError>

This method goes through the motions of load, but only returns the metadata.

If dry_load returns an error, load must return the same error, but not vice-versa. Concretely, load could return deserialization or I/O errors that dry_load cannot predict.